+++ src/config.rs
@@ -1,15 +1,8 @@
-// fast-down task config module
-// 模块名称: config
-// 职责范围: 任务配置管理，读取保存目录下的fast-down.yaml文件
-// 期望实现计划: 支持download一级目录下的URL任务配置，dir为相对路径
-// 已实现功能: YAML配置解析、URL任务提取、相对路径处理、示例配置生成
-// 使用依赖: serde, serde_yaml, std::path, std::collections
-// 主要接口: TaskConfig::load_from_file, TaskConfig::get_tasks, TaskConfig::load_from_save_dir, create_example_config
-// 注意事项: 只处理任务相关配置，不干扰原有配置系统，支持并发任务处理
-
+use color_eyre::Result;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
+use tokio::fs;
 
 /// 任务配置结构
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -36,23 +29,22 @@ pub struct TaskSettings {
 
 impl TaskConfig {
     /// 从指定路径加载任务配置
-    pub fn load_from_file<P: AsRef<Path>>(path: P) -> Result<Self, Box<dyn std::error::Error>> {
-        let content = std::fs::read_to_string(path)?;
+    pub async fn load_from_file<P: AsRef<Path>>(path: P) -> Result<Self> {
+        let content = fs::read_to_string(path).await?;
         let config: TaskConfig = serde_yaml::from_str(&content)?;
         Ok(config)
     }
 
     /// 获取所有任务
-    pub fn get_tasks(&self, base_dir: &Path) -> Vec<TaskItem> {
+    pub fn get_tasks<P: AsRef<Path>>(&self, base_dir: P) -> Vec<TaskItem> {
         self.download
             .iter()
             .map(|(url, settings)| {
                 let save_path = if let Some(dir) = &settings.dir {
-                    base_dir.join(dir)
+                    base_dir.as_ref().join(dir)
                 } else {
-                    base_dir.to_path_buf()
+                    base_dir.as_ref().to_path_buf()
                 };
-
                 TaskItem {
                     url: url.clone(),
                     save_path,
@@ -60,15 +52,6 @@ impl TaskConfig {
                 }
             })
             .collect()
-    }
-
-    /// 从保存目录加载fast-down.yaml
-    pub fn load_from_save_dir(save_dir: &Path) -> Result<Self, Box<dyn std::error::Error>> {
-        let config_path = save_dir.join("fast-down.yaml");
-        if !config_path.exists() {
-            return Err("fast-down.yaml not found".into());
-        }
-        Self::load_from_file(config_path)
     }
 }
